+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Change log
 ----------------------
 
-Version 3.1-SNAPSHOT
+Version 3.1
 -------------
+
+CHANGED:
+
+- replaced all dependecies from 'javax.*' with 'jakarta.*'
 
 Version 3
 -------------
@@ -18,7 +22,6 @@ CHANGED:
 - removed abstract configuration class for swagger
 - update of dependency spring to new version 6.0.3
 - update of test dependency mockito-core to new version 4.10.0
-- replaced all dependecies from 'javax.*' with 'jakarta.*'
 
 Version 2.1
 -------------